requirements.txt
Flask==2.2.5
numpy==1.24.3
pandas==1.3.3
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.3.0
imbalanced-learn==0.8.1
plotly==5.3.1
opencv-python==4.8.0
Pillow==9.4.0
WTForms==2.3.3
Flask-WTF==0.15.1
Flask-Bootstrap==3.3.7.1
Flask-Assets==2.0
tensorflow==2.12.0
xgboost==1.7.3
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.5